support setup.py
