--- layers/__init__.py
@@ -1,6 +1,0 @@
-from layers.conv_layers import *
-from layers.dimension_utils import *
-from layers.common_layers import *
-from layers.activations_layers import *
-from layers.mathematics_layers import *
-from layers.deformation_layers import *
