+++ onnx2tflite/__init__.py
@@ -1,0 +1,3 @@
+__VERSION__ = "2.0"
+
+from converter import onnx_converter
