+++ onnx2tflite/layers/activations_layers.py
@@ -2,8 +2,8 @@ import numpy as np
 import tensorflow as tf
 from tensorflow import keras
 
-from layers.dimension_utils import channel_to_last_dimension, tensor_NCD_to_NDC_format
-from utils.op_registry import OPERATOR
+from .dimension_utils import channel_to_last_dimension, tensor_NCD_to_NDC_format
+from onnx2tflite.utils.op_registry import OPERATOR
 
 @OPERATOR.register_operator("Relu")
 class TFRelu():
