+++ onnx2tflite/layers/common_layers.py
@@ -4,8 +4,8 @@ import numpy as np
 import tensorflow as tf
 from tensorflow import keras
 
-from utils.op_registry import OPERATOR
-from layers.dimension_utils import intfloat_to_list, channel_to_last_dimension
+from onnx2tflite.utils.op_registry import OPERATOR
+from .dimension_utils import intfloat_to_list, channel_to_last_dimension
 
 LOG = logging.getLogger("common_layers :")
