+++ onnx2tflite/layers/mathematics_layers.py
@@ -2,8 +2,8 @@ import logging
 import numpy as np
 import tensorflow as tf
 
-from utils.op_registry import OPERATOR
-from layers import dimension_utils
+from onnx2tflite.utils.op_registry import OPERATOR
+from . import dimension_utils
 
 LOG = logging.getLogger("calculations_layers :")
