--- utils/__init__.py
@@ -1,5 +1,0 @@
-from utils.output_check import get_elements_error
-from utils.onnx_loader import load_onnx_modelproto
-from utils.builder import keras_builder, tflite_builder
-
-__all__ = ['load_onnx_modelproto', 'keras_builder', 'tflite_builder', 'get_elements_error']
